Fix determining gem version from latest git tag
It was showing an error:

```
fatal: no tag exactly matches '3467704e40171157184da6f237196c9f4c1b494d'
```

and the version of the gem in the `Gemfile.lock` was being adjusted to 0!

This is different from all (?) of our other gems, which have a version constant defined in a Ruby file. Fixing this was easier than reworking it. I take it this issue is for changing this: https://github.com/greensync/omniauth-dex-energy/issues/9